requirements.txt
streamlit
onnxruntime
Pillow
numpy


